+++ src/Promitor.Agents.ResourceDiscovery/Controllers/v1/ResourceDiscoveryGroupsV1Controller.cs
@@ -10,14 +10,14 @@ namespace Promitor.Agents.ResourceDiscovery.Controllers.v1
     /// </summary>
     [ApiController]
     [Route("api/v1/resources/groups")]
-    public class ResourceDiscoveryGroupsController : ControllerBase
+    public class ResourceDiscoveryGroupsV1Controller : ControllerBase
     {
         private readonly IOptionsMonitor<ResourceDeclaration> _resourceDeclarationMonitor;
 
         /// <summary>
-        ///     Initializes a new instance of the <see cref="ResourceDiscoveryGroupsController" /> class.
+        ///     Initializes a new instance of the <see cref="ResourceDiscoveryGroupsV1Controller" /> class.
         /// </summary>
-        public ResourceDiscoveryGroupsController(IOptionsMonitor<ResourceDeclaration> resourceDeclarationMonitor)
+        public ResourceDiscoveryGroupsV1Controller(IOptionsMonitor<ResourceDeclaration> resourceDeclarationMonitor)
         {
             Guard.NotNull(resourceDeclarationMonitor, nameof(resourceDeclarationMonitor));
