+++ src/Promitor.Agents.ResourceDiscovery/Controllers/v1/SystemV1Controller.cs
@@ -7,7 +7,7 @@ using Swashbuckle.AspNetCore.Annotations;
 namespace Promitor.Agents.ResourceDiscovery.Controllers.v1
 {
     [Route("api/v1/system")]
-    public class SystemController : Controller
+    public class SystemV1Controller : Controller
     {
         /// <summary>
         ///     Get System Info
